Cleanup and README updates
