+++ examples/es-clusters.yaml
@@ -32,6 +32,8 @@ metadata:
 
     # Optional if regions are declared but the es-endpoint link is
     # a link to just one cluster you can specify the region with this label
+    # Every region becomes a column in the ES table.
+    # This value is not used if the es-endpoint contains a {region} variable
     # - elasticsearch-plugin/region: gew1
 
     # The endpoint (pattern) to the cluster
@@ -43,6 +45,7 @@ metadata:
     # HTTP Example
     # - elasticsearch-plugin/es-endpoint: "http://my-cluster.{region}.domain.com:9200"
     
+    # [Optionial] Links to Cerebro and Kibana
     # Variables:
     #   {es-endpoint}           - Full value of es-endpoint (example "https://my-cluster.region1.domain.com:9200")
     #   {es-endpoint.hostname}  - Hostname from the es-endpoint (example "my-cluster.region1.domain.com")
@@ -112,12 +115,12 @@ apiVersion: backstage.io/v1alpha1
 kind: Component
 metadata:
   name: it2
-  description: Sample ES cluster discovered with SRV only on one region
+  description: Single ES cluster discovered with SRV only on one region
   label:
     - elasticsearch-plugin/region: gew1
-      
+
     - elasticsearch-plugin/es-endpoint: "srv:_it2-master._http.services.gew1.mydomain.net"
-    
+
     - elasticsearch-plugin/kibana-endpoint: "{es-endpoint.scheme}://{es-endpoint.hostname}:5601"
     - elasticsearch-plugin/cerebro-endpoint: "http://cerebro.eu-region.mydomain.net:9000/#!/overview?host={es-endpoint}"
   tags:
