+++ plugins/elasticsearch-clusters-backend/src/service/config.ts
@@ -1,8 +1,7 @@
 
 export interface EsConfig {
-    cerebro_link: string | undefined;
-    kibana_link: string | undefined;
+   
     regions: string[] | undefined;
-    cluster_links: string[] | undefined;
+    cacheControl: string | undefined;
   }
